Reorganize maps and layers and fix some CSS
