+++ src/app/app.component.html
@@ -1,4 +1,4 @@
 <igo-header></igo-header>
-<section id="router-body">
+<main id="router-body">
   <router-outlet></router-outlet>
-</section>
+</main>
