+++ src/app/map/map.component.styl
@@ -1,4 +1,4 @@
-:host, .ngmap-map-target {
+:host, .igo-map-target {
   width: 100%;
   height: 100%;
 }
