+++ src/app/map/shared/layers/layer-osm.interface.ts
@@ -1,0 +1,6 @@
+import { LayerOptions } from './layer.interface';
+
+export interface OSMLayerOptions extends LayerOptions {
+  source?: olx.source.OSMOptions;
+  view?: olx.layer.TileOptions;
+}
