+++ src/app/map/shared/layers/layer-osm.model.ts
@@ -1,9 +1,5 @@
-import { Layer, LayerOptions } from './layer.model';
-
-export interface OSMLayerOptions extends LayerOptions {
-  source?: olx.source.OSMOptions;
-  view?: olx.layer.TileOptions;
-}
+import { Layer } from './layer.model';
+import { OSMLayerOptions } from './layer-osm.interface';
 
 export class OSMLayer extends Layer {
