+++ src/app/map/shared/layers/layer-tile.interface.ts
@@ -1,0 +1,6 @@
+import { LayerOptions } from './layer.interface';
+
+export interface TileLayerOptions extends LayerOptions {
+  source?: olx.source.TileImageOptions;
+  view?: olx.layer.TileOptions;
+}
