+++ src/app/map/shared/layers/layer.interface.ts
@@ -1,0 +1,4 @@
+export interface LayerOptions {
+  name: string;
+  type: string;
+}
