+++ src/app/map/shared/layers/layer.model.ts
@@ -1,7 +1,4 @@
-export interface LayerOptions {
-  name: string;
-  type: string;
-}
+import { LayerOptions } from './layer.interface';
 
 export abstract class Layer {
