+++ src/app/map/shared/map.interface.ts
@@ -1,0 +1,3 @@
+export interface MapOptions {
+  view: olx.ViewOptions;
+}
