+++ src/app/map/shared/map.model.ts
@@ -1,8 +1,6 @@
-import { Layer, LayerOptions } from './layers/layer.model';
-
-export interface MapOptions {
-  view: olx.ViewOptions;
-}
+import { MapOptions } from './map.interface';
+import { LayerOptions } from './layers/layer.interface';
+import { Layer } from './layers/layer.model';
 
 // The name is NgMap because Map exists already
 export class NgMap {
