+++ src/app/navigator-page/navigator-page.component.html
@@ -1,4 +1,4 @@
 <igo-map
-  [view]="config.map.view"
-  [layers]="config.map.layers">
+  [options]="context.map"
+  [layers]="context.layers">
 </igo-map>
