+++ src/css/styles.styl
@@ -7,6 +7,11 @@ div {
   padding: 0;
 }
 
+html, body {
+  width: 100%;
+  height: 100%;
+}
+
 /* Box-sizing border-box */
 * {
   box-sizing: border-box;
